requirements.txt
pyserial
websockets
asyncio
tzlocal
